backend/requirements.txt
accelerate==1.12.0
addict==2.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-cors==0.8.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
arel==0.4.0
async-asgi-testclient==1.4.11
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.2
bidict==0.23.1
bleach==6.3.0
blis==1.3.3
cacheout==0.14.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.2.1
cloudpathlib==0.23.0
colorful==0.5.8
confection==0.1.5
contourpy==1.3.2
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
cymem==2.0.13
dataframe_image==0.2.7
dataset-tools==0.1.9
debugpy==1.8.17
defusedxml==0.7.1
distinctipy==1.3.4
distlib==0.4.0
einops==0.8.1
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.119.1
fastjsonschema==2.21.2
ffmpeg-python==0.2.0
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
gdown==5.2.0
geojson==3.2.0
gitdb==4.0.12
GitPython==3.1.45
giturlparse==0.14.0
google-api-core==2.28.1
google-auth==2.43.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httptools==0.7.1
httpx==0.27.2
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
imutils==0.5.4
inflect==7.5.0
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
lazy_loader==0.4
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.1.4
mkl_fft==1.3.8
mkl_random==1.2.4
mkl-service==2.4.0
mmcv-full==1.7.1
mmengine==0.10.7
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
murmurhash==1.0.15
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
numerize==0.12
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.11.0.86
opentelemetry-api==1.38.0
opentelemetry-exporter-prometheus==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pandas==1.5.2
pandocfilters==1.5.1
peft==0.18.0
pillow==10.4.0
pip==25.3
platformdirs==4.5.0
playwright==1.56.0
preshed==3.0.12
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
PTable==0.9.2
py-spy==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pydantic==2.12.3
pydantic_core==2.41.4
pydicom==2.4.4
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
Pympler==1.0.1
pynrrd==0.4.3
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.12
pytz==2025.2
PyYAML==6.0.3
ray==2.52.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==0.10.1
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
setuptools==80.9.0
shapely==2.1.2
SimpleITK==2.4.1
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
soupsieve==2.8
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.47.3
stringcase==1.2.0
supervisely==6.73.475
sympy==1.14.0
termcolor==3.2.0
thinc==8.3.10
threadpoolctl==3.5.0
tifffile==2025.5.10
timm==1.0.22
tinycss2==1.4.0
titlecase==2.4.1
tokenizers==0.22.1
tomli==2.3.0
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
trimesh==4.5.0
triton==3.5.1
typeguard==4.4.4
typer-slim==0.20.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==1.26.15
uvicorn==0.38.0
uvloop==0.22.1
varname==0.15.0
virtualenv==20.35.4
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
webencodings==0.5.1
websockets==13.1
wheel==0.45.1
wrapt==2.0.1
xmltodict==0.13.0
yapf==0.43.0
yarl==1.22.0
zipp==3.23.0
zstd==1.5.7.2
